backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
jinja2==3.1.2
pandas==2.1.4
numpy==1.25.2
aiofiles==23.2.1
email-validator==2.1.0
google-generativeai==0.3.2
